Move BRS_X04 to BRS_X03.OrchestrationDescriptionBreakingChanges
