+++ source/ProcessManager.Example.Orchestrations.Tests/Integration/Monitor/HealthCheckEndpointTests.cs
@@ -15,7 +15,7 @@
 using System.Net;
 using Energinet.DataHub.ProcessManager.Core.Domain.OrchestrationDescription;
 using Energinet.DataHub.ProcessManager.Core.Infrastructure.Extensions.Options;
-using Energinet.DataHub.ProcessManager.Example.Orchestrations.Processes.BRS_X04_OrchestrationDescriptionBreakingChanges;
+using Energinet.DataHub.ProcessManager.Example.Orchestrations.Processes.BRS_X03.OrchestrationDescriptionBreakingChanges;
 using Energinet.DataHub.ProcessManager.Example.Orchestrations.Tests.Fixtures;
 using Energinet.DataHub.ProcessManager.Shared.Tests.Fixtures.Extensions;
 using FluentAssertions;
